Select lowest rendition available
